--- Uno/src/game/A_MainRun.java
@@ -1,81 +1,0 @@
-package game;
-
-import java.io.File;
-import javafx.application.Application;
-import javafx.fxml.*;
-import javafx.scene.Group;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.Pane;
-import javafx.scene.layout.StackPane;
-import javafx.scene.media.*;
-import javafx.scene.paint.Color;
-import javafx.stage.Stage;
-
-public class A_MainRun extends Application {
-
-    private String DATA_PATH ;//= "../Engo/src/";
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage stage) throws Exception {
-        //StackPane pane = new StackPane();
-        Parent root = FXMLLoader.load(getClass().getResource("Menu.fxml"));
-//        FXMLLoader loader = new FXMLLoader(new File("fullpath").toURI().toURL());
-//        Parent root = loader.load();
-
-        Pane root2 = new Pane();
-        ImageView background = new ImageView(new Image("img/logo.png", 1080, 720, false, true));
-        root2.getChildren().add(background);
-
-        //Group root = new Group();
-        Scene scene = new Scene(root, 1080, 720);
-
-        stage.setTitle("ENGO!");
-
-        //add pic icon
-        Image icon = new Image("img/engoLogo.png");
-        stage.getIcons().add(icon);
-
-        //add pic menu
-        Image logo = new Image("img/logonobg.png");
-//        Image logo = new Image(new File("img/logonobg.png").toURI().toURL()); -> onn test can not resolve this QAQ
-        ImageView imageView = new ImageView(logo);
-
-        // --------------- ADD SOUND ----------------------
-//        String musicFile =   "../Uno/src/sound/howtoplay.mp3";
-//
-//        File file = new File(musicFile);
-//        System.out.println("isfile exists : " + file.exists());
-//        System.out.println("isfile : " + file.isFile());
-//
-//        Media sound = new Media(new File(musicFile).toURI().toString());
-//        MediaPlayer mediaPlayer = new MediaPlayer(sound);
-//        mediaPlayer.play();
-//--------------------
-//        Media media = new Media(new File(musicFile).toURI().toString());
-//         
-//        //Instantiating MediaPlayer class   
-//        MediaPlayer mediaPlayer = new MediaPlayer(media);  
-//          
-//        //by setting this property to true, the audio will be played   
-//        mediaPlayer.setAutoPlay(true);  
-
-        /*imageView.setX(20);
-        imageView.setY(-30);
-        imageView.setScaleX(0.5);
-        imageView.setScaleY(0.5);
-        root.getChildren().add(imageView);*/
-        //scene.getStylesheets().add("http://fonts.googleapis.com/css?family=Gafata%22");
-        scene.getStylesheets().add(getClass().getResource("UnoStyle.css").toExternalForm());
-        stage.setScene(scene);
-        stage.setResizable(false);
-        stage.show();
-    }
-
-}
