--- Uno/src/game/BotClass.java
@@ -1,16 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package game;
-
-/**
- *
- * @author chura
- */
-public class BotClass {
-    
-    
-//    if(get)   
-}
