--- Uno/src/game/Direct.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="400.0" prefWidth="600.0" styleClass="direct" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="uno.DirectController">
-   <children>
-      <VBox alignment="CENTER" layoutX="10.0" layoutY="10.0" prefHeight="720.0" prefWidth="1080.0" styleClass="direct">
-         <children>
-            <Label text="Directors" textFill="WHITE">
-               <font>
-                  <Font size="39.0" />
-               </font>
-            </Label>
-            <HBox alignment="CENTER" prefHeight="100.0" prefWidth="1080.0">
-               <children>
-                  <Button mnemonicParsing="false" onAction="#actionBack" text="Back">
-                     <font>
-                        <Font size="35.0" />
-                     </font>
-                  </Button>
-               </children>
-            </HBox>
-         </children>
-      </VBox>
-   </children>
-</AnchorPane>
