--- Uno/src/game/Play.fxml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="720.0" prefWidth="1080.0" styleClass="play" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="uno.PlayController">
-   <children>
-      <VBox alignment="CENTER" prefHeight="720.0" prefWidth="1080.0" AnchorPane.bottomAnchor="100.0" AnchorPane.topAnchor="100.0">
-         <children>
-            <Label alignment="CENTER" contentDisplay="CENTER" prefHeight="142.0" prefWidth="1080.0" text="Name" textAlignment="CENTER" textFill="WHITE">
-               <font>
-                  <Font size="37.0" />
-               </font>
-            </Label>
-            <TextField fx:id="name" alignment="CENTER" prefHeight="65.0" prefWidth="1018.0" promptText="Please enter your name!">
-               <VBox.margin>
-                  <Insets left="150.0" right="150.0" />
-               </VBox.margin>
-            </TextField>
-            <HBox alignment="CENTER" prefHeight="164.0" prefWidth="1080.0">
-               <children>
-                  <Button fx:id="setName" mnemonicParsing="false" onAction="#actionSetName" text="Play">
-                     <font>
-                        <Font size="21.0" />
-                     </font>
-                  </Button>
-                  <Button fx:id="home" mnemonicParsing="false" onAction="#actionBack" text="Back">
-                     <font>
-                        <Font size="21.0" />
-                     </font>
-                  </Button>
-               </children>
-            </HBox>
-         </children>
-      </VBox>
-   </children>
-</AnchorPane>
