--- Uno/src/game/UnoStyle.css
@@ -1,99 +1,0 @@
-/*
-To change this license header, choose License Headers in Project Properties.
-To change this template file, choose Tools | Templates
-and open the template in the editor.
-*/
-/* 
-    Created on : Apr 24, 2021, 3:14:05 AM
-    Author     : chura
-*/
-*{
-    -fx-font-family: "Cooper Black";
-    /*src: url('CooperBlack.ttf');*/
-    /*-fx-font-size: 20;*/
-}
-
-.root{
-    -fx-background-color: linear-gradient(violet, white);
-}
-
-.menu{
-    /*-fx-background-image: url(../logo.png);*/
-}
-
-.howto .label,
-.direct .label,
-.play .label,
-.rank .label{
-    -fx-text-alignment: center;
-    -fx-font-size: 50;
-}
-.text-field {
-    -fx-font-size: 30;
-    -fx-text-fill:  #cc00cc;  
-    -fx-background-radius: 50;
-    -fx-background-insets: 0 -1 -1 -1, 0 0 0 0, 0 -1 3 -1;
-}
-
-.text-field:focused{
-    -fx-effect: dropshadow( one-pass-box , rgba(120,120,120,0.8) , 30, 0.0 , 0 , 1);
-    
-}
-
-.rank .text-field,
-.game .text-field{
-   -fx-background-color: transparent;
-   -fx-text-alignment: center;  
-   -fx-text-fill: #009966;
-}
-
-/*.button,*/
-.menu .button,
-.howto .button,
-.direct .button,
-.play .button,
-.rank .button,
-.game .button{
-    -fx-background-color: #ff9999;
-    -fx-background-radius: 50;
-    /*-fx-background-insets: 10;*/
-    /*-fx-background-image: url(../img/btn.png);*/
-    -fx-text-fill: blueviolet;
-    -fx-font-weight: bold;      
-    -fx-font-size: 20;
-}
-
-.button:hover{
-    -fx-background-color: fuchsia; 
-    -fx-background-radius: 50;
-    -fx-text-fill: greenyellow;
-    -fx-rotate: 5;
-    -fx-effect: dropshadow( one-pass-box , rgba(0,0,0,0.8) , 30, 0.0 , 0 , 1);
-}
-
-.button:pressed{
-    /*-fx-background-color: white;*/
-    /*-fx-rotate: 10;*/
-}
-
-/*#playNow{
-    fx-color:cyan;
-    color: green;
-    -fx-background-color: black;
-    -fx-text-fill: aquamarine;
-}*/
-
-.progessBar{
-    -fx-width: 100;
-}
-
-.progress-bar > .bar {
-    -fx-background-color: linear-gradient(
-        from 0px .75em to .75em 0px,
-        repeat,
-        -fx-accent 0%,
-        -fx-accent 49%,
-        derive(-fx-accent, 30%) 50%,
-        derive(-fx-accent, 30%) 99%
-    );
-}
