+++ Uno/src/gameengo/MAIN.fxml
@@ -3,7 +3,6 @@
 <?import javafx.geometry.Insets?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -11,7 +10,7 @@
 <?import javafx.scene.shape.Rectangle?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane id="AnchorPane" prefHeight="720.0" prefWidth="1080.0" styleClass="game" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gameengo.MAINController">
+<AnchorPane id="AnchorPane" prefHeight="720.0" prefWidth="1080.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gameengo.MAINController">
    <children>
       <AnchorPane prefHeight="724.0" prefWidth="1079.0">
          <children>
@@ -30,7 +29,6 @@
                   <Font size="23.0" />
                </font>
             </Label>
-            <TextField fx:id="name" layoutX="5.0" layoutY="650.0" prefHeight="55.0" prefWidth="258.0" />
             <ImageView fx:id="StopPic" fitHeight="328.0" fitWidth="434.0" layoutX="-89.0" layoutY="199.0" opacity="0.0" pickOnBounds="true" preserveRatio="true">
                <image>
                   <Image url="@../pic/prohibited.png" />
