+++ Uno/src/uno/MenuController.java
@@ -1,0 +1,87 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package uno;
+
+import java.io.IOException;
+import java.net.URL;
+import java.util.ResourceBundle;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.fxml.Initializable;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.scene.image.Image;
+import javafx.stage.Stage;
+
+
+/**
+ * FXML Controller class
+ *
+ * @author chura
+ */
+public class MenuController implements Initializable {
+
+    @FXML
+    private Button play;
+    @FXML
+    private Button howto;
+    @FXML
+    private Button direct;
+    @FXML
+    private Button exit;
+
+    private Stage stage;
+    private Scene scene;
+    private Parent root;
+    Image icon = new Image("img/logo.png");
+
+    /**
+     * Initializes the controller class.
+     */
+    @Override
+    public void initialize(URL url, ResourceBundle rb) {
+        // TODO
+    }
+
+    @FXML
+    private void actionPlay(ActionEvent event) throws IOException {
+        root = FXMLLoader.load(getClass().getResource("Play.fxml"));
+        stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
+        scene = new Scene(root);
+        scene.getStylesheets().add(getClass().getResource("UnoStyle.css").toExternalForm());
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    @FXML
+    private void actionHowto(ActionEvent event) throws IOException {
+        root = FXMLLoader.load(getClass().getResource("Howto.fxml"));
+        stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
+        scene = new Scene(root);
+        scene.getStylesheets().add(getClass().getResource("UnoStyle.css").toExternalForm());
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    @FXML
+    private void actionDirect(ActionEvent event) throws IOException {
+        root = FXMLLoader.load(getClass().getResource("Direct.fxml"));
+        stage = (Stage) ((Node) event.getSource()).getScene().getWindow();
+        scene = new Scene(root);
+        scene.getStylesheets().add(getClass().getResource("UnoStyle.css").toExternalForm());
+        stage.setScene(scene);
+        stage.show();
+    }
+
+    @FXML
+    private void actionExit(ActionEvent event) {
+        ((Stage) (((Button) event.getSource()).getScene().getWindow())).close();
+    }
+
+}
