+++ Uno/src/uno/UnoStyle.css
@@ -1,0 +1,72 @@
+/*
+To change this license header, choose License Headers in Project Properties.
+To change this template file, choose Tools | Templates
+and open the template in the editor.
+*/
+/* 
+    Created on : Apr 24, 2021, 3:14:05 AM
+    Author     : chura
+*/
+*{
+    -fx-font-family: "Cooper Black";
+    /*src: url('CooperBlack.ttf');*/
+}
+
+.root{
+    -fx-background-color: violet;
+}
+.menu{
+    /*-fx-background-image: url(../logo.png);*/
+}
+
+.menu .button,
+.howto .button,
+.direct .button,
+.play .button {
+    -fx-background-color: plum;
+    -fx-background-radius: 50;
+    /*-fx-background-insets: 10;*/
+    /*-fx-background-image: url(../img/btn.png);*/
+     
+    -fx-text-fill: blueviolet;
+    -fx-font-weight: bold;      
+    -fx-font-size: 20;
+}
+
+.button:hover{
+    -fx-background-color: fuchsia;
+    -fx-background-radius: 50;
+    -fx-text-fill: greenyellow;
+    /*-fx-text-fill: linear-gradient(white, #d0d0d0);*/
+    -fx-animation: faded in;
+    -fx-transition: 2;
+    -fx-rotate: 7;
+    -fx-effect: dropshadow( one-pass-box , rgba(0,0,0,0.8) , 30, 0.0 , 0 , 1);
+}
+
+.button:pressed{
+    /*-fx-background-color: white;*/
+    /*-fx-rotate: 10;*/
+}
+
+/*#playNow{
+    fx-color:cyan;
+    color: green;
+    -fx-background-color: black;
+    -fx-text-fill: aquamarine;
+}*/
+
+.progessBar{
+    -fx-width: 100;
+}
+
+.progress-bar > .bar {
+    -fx-background-color: linear-gradient(
+        from 0px .75em to .75em 0px,
+        repeat,
+        -fx-accent 0%,
+        -fx-accent 49%,
+        derive(-fx-accent, 30%) 50%,
+        derive(-fx-accent, 30%) 99%
+    );
+}
